+++ NEWS.md
@@ -19,6 +19,7 @@ Version 0.17.1b2 - dev
     by a block reference
   - `source_block_reference` - the immediate source block reference (`INSERT`), 
     which created the virtual entity, otherwise ``None``
+- NEW: `ezdxf.tools.text_size` module to measure `TEXT` and `MTEXT` entity dimensions
 - BUGFIX: invalid XDATA processing in `XData.safe_init()` 
 - BUGFIX: group code 1003 is valid in XDATA section 
 - BUGFIX: fix loading error of `DIMSTYLE` attribute `dimtxsty`